Pods 2.3.19 Beta 1

+++ includes/general.php
@@ -1115,7 +1115,13 @@ function pods_field ( $pod, $id = false, $name = null, $single = false ) {
         $id = get_the_ID();
     }
 
-    return pods( $pod, $id )->field( $name, $single );
+    $pod = pods( $pod, $id );
+
+	if ( is_object( $pod ) ) {
+		return $pod->field( $name, $single );
+	}
+
+	return null;
 }
 
 /**
@@ -1140,7 +1146,13 @@ function pods_field_display ( $pod, $id = false, $name = null, $single = false )
         $id = get_the_ID();
     }
 
-    return pods( $pod, $id )->display( $name, $single );
+    $pod = pods( $pod, $id );
+
+	if ( is_object( $pod ) ) {
+		return $pod->display( $name, $single );
+	}
+
+	return null;
 }
 
 /**
@@ -1627,6 +1639,7 @@ function pods_no_conflict_check ( $object_type = 'post' ) {
  * @since 2.0
  */
 function pods_no_conflict_on ( $object_type = 'post', $object = null ) {
+
     if ( 'post_type' == $object_type )
         $object_type = 'post';
     elseif ( 'term' == $object_type )
@@ -1638,22 +1651,29 @@ function pods_no_conflict_on ( $object_type = 'post', $object = null ) {
     if ( !is_object( PodsInit::$meta ) )
         return false;
 
-    $no_conflict = array();
+    $no_conflict = array(
+		'filter' => array()
+	);
 
     // Filters = Usually get/update/delete meta functions
     // Actions = Usually insert/update/save/delete object functions
     if ( 'post' == $object_type ) {
-        $no_conflict[ 'filter' ] = array(
-            array( 'get_post_metadata', array( PodsInit::$meta, 'get_post_meta' ), 10, 4 ),
-        );
+		if ( apply_filters( 'pods_meta_handler', true, 'post' ) ) {
+            // Handle *_post_meta
+			if ( apply_filters( 'pods_meta_handler_get', true, 'post' ) ) {
+				$no_conflict[ 'filter' ] = array(
+					array( 'get_post_metadata', array( PodsInit::$meta, 'get_post_meta' ), 10, 4 ),
+				);
+			}
 
-        if ( !pods_tableless() ) {
-            $no_conflict[ 'filter' ] = array_merge( $no_conflict[ 'filter' ], array(
-                array( 'add_post_metadata', array( PodsInit::$meta, 'add_post_meta' ), 10, 5 ),
-                array( 'update_post_metadata', array( PodsInit::$meta, 'update_post_meta' ), 10, 5 ),
-                array( 'delete_post_metadata', array( PodsInit::$meta, 'delete_post_meta' ), 10, 5 )
-            ) );
-        }
+			if ( !pods_tableless() ) {
+				$no_conflict[ 'filter' ] = array_merge( $no_conflict[ 'filter' ], array(
+					array( 'add_post_metadata', array( PodsInit::$meta, 'add_post_meta' ), 10, 5 ),
+					array( 'update_post_metadata', array( PodsInit::$meta, 'update_post_meta' ), 10, 5 ),
+					array( 'delete_post_metadata', array( PodsInit::$meta, 'delete_post_meta' ), 10, 5 )
+				) );
+			}
+		}
 
         $no_conflict[ 'action' ] = array(
             array( 'transition_post_status', array( PodsInit::$meta, 'save_post_detect_new' ), 10, 3 ),
@@ -1661,41 +1681,59 @@ function pods_no_conflict_on ( $object_type = 'post', $object = null ) {
         );
     }
     elseif ( 'taxonomy' == $object_type ) {
-        $no_conflict[ 'filter' ] = array();
+		if ( apply_filters( 'pods_meta_handler', true, 'term' ) ) {
+            // Handle *_term_meta
+			if ( apply_filters( 'pods_meta_handler_get', true, 'term' ) ) {
+        		$no_conflict[ 'filter' ] = array();
+			}
 
-        $no_conflict[ 'action' ] = array(
-            array( 'edit_term', array( PodsInit::$meta, 'save_taxonomy' ), 10, 3 ),
-            array( 'create_term', array( PodsInit::$meta, 'save_taxonomy' ), 10, 3 )
-        );
+			$no_conflict[ 'action' ] = array(
+				array( 'edit_term', array( PodsInit::$meta, 'save_taxonomy' ), 10, 3 ),
+				array( 'create_term', array( PodsInit::$meta, 'save_taxonomy' ), 10, 3 )
+			);
+		}
     }
     elseif ( 'media' == $object_type ) {
-        $no_conflict[ 'filter' ] = array(
-            array( 'wp_update_attachment_metadata', array( PodsInit::$meta, 'save_media' ), 10, 2 ),
-            array( 'get_post_metadata', array( PodsInit::$meta, 'get_post_meta' ), 10, 4 )
-        );
+		$no_conflict[ 'filter' ] = array(
+			array( 'wp_update_attachment_metadata', array( PodsInit::$meta, 'save_media' ), 10, 2 )
+		);
 
-        if ( !pods_tableless() ) {
-            $no_conflict[ 'filter' ] = array_merge( $no_conflict[ 'filter' ], array(
-                array( 'add_post_metadata', array( PodsInit::$meta, 'add_post_meta' ), 10, 5 ),
-                array( 'update_post_metadata', array( PodsInit::$meta, 'update_post_meta' ), 10, 5 ),
-                array( 'delete_post_metadata', array( PodsInit::$meta, 'delete_post_meta' ), 10, 5 )
-            ) );
-        }
+		if ( apply_filters( 'pods_meta_handler', true, 'post' ) ) {
+            // Handle *_post_meta
+			if ( apply_filters( 'pods_meta_handler_get', true, 'post' ) ) {
+				$no_conflict[ 'filter' ] = array_merge( $no_conflict[ 'filter' ], array(
+					array( 'get_post_metadata', array( PodsInit::$meta, 'get_post_meta' ), 10, 4 )
+				) );
+			}
 
-        $no_conflict[ 'action' ] = array();
+			if ( !pods_tableless() ) {
+				$no_conflict[ 'filter' ] = array_merge( $no_conflict[ 'filter' ], array(
+					array( 'add_post_metadata', array( PodsInit::$meta, 'add_post_meta' ), 10, 5 ),
+					array( 'update_post_metadata', array( PodsInit::$meta, 'update_post_meta' ), 10, 5 ),
+					array( 'delete_post_metadata', array( PodsInit::$meta, 'delete_post_meta' ), 10, 5 )
+				) );
+			}
+
+			$no_conflict[ 'action' ] = array();
+		}
     }
     elseif ( 'user' == $object_type ) {
-        $no_conflict[ 'filter' ] = array(
-            array( 'get_user_metadata', array( PodsInit::$meta, 'get_user_meta' ), 10, 4 ),
-        );
+		if ( apply_filters( 'pods_meta_handler', true, 'user' ) ) {
+            // Handle *_term_meta
+			if ( apply_filters( 'pods_meta_handler_get', true, 'user' ) ) {
+				$no_conflict[ 'filter' ] = array(
+					array( 'get_user_metadata', array( PodsInit::$meta, 'get_user_meta' ), 10, 4 ),
+				);
+			}
 
-        if ( !pods_tableless() ) {
-            $no_conflict[ 'filter' ] = array_merge( $no_conflict[ 'filter' ], array(
-                array( 'add_user_metadata', array( PodsInit::$meta, 'add_user_meta' ), 10, 5 ),
-                array( 'update_user_metadata', array( PodsInit::$meta, 'update_user_meta' ), 10, 5 ),
-                array( 'delete_user_metadata', array( PodsInit::$meta, 'delete_user_meta' ), 10, 5 )
-            ) );
-        }
+			if ( !pods_tableless() ) {
+				$no_conflict[ 'filter' ] = array_merge( $no_conflict[ 'filter' ], array(
+					array( 'add_user_metadata', array( PodsInit::$meta, 'add_user_meta' ), 10, 5 ),
+					array( 'update_user_metadata', array( PodsInit::$meta, 'update_user_meta' ), 10, 5 ),
+					array( 'delete_user_metadata', array( PodsInit::$meta, 'delete_user_meta' ), 10, 5 )
+				) );
+			}
+		}
 
         $no_conflict[ 'action' ] = array(
             //array( 'user_register', array( PodsInit::$meta, 'save_user' ) ),
@@ -1703,17 +1741,22 @@ function pods_no_conflict_on ( $object_type = 'post', $object = null ) {
         );
     }
     elseif ( 'comment' == $object_type ) {
-        $no_conflict[ 'filter' ] = array(
-            array( 'get_comment_metadata', array( PodsInit::$meta, 'get_comment_meta' ), 10, 4 ),
-        );
+		if ( apply_filters( 'pods_meta_handler', true, 'comment' ) ) {
+            // Handle *_term_meta
+			if ( apply_filters( 'pods_meta_handler_get', true, 'comment' ) ) {
+				$no_conflict[ 'filter' ] = array(
+					array( 'get_comment_metadata', array( PodsInit::$meta, 'get_comment_meta' ), 10, 4 ),
+				);
+			}
 
-        if ( !pods_tableless() ) {
-            $no_conflict[ 'filter' ] = array_merge( $no_conflict[ 'filter' ], array(
-                array( 'add_comment_metadata', array( PodsInit::$meta, 'add_comment_meta' ), 10, 5 ),
-                array( 'update_comment_metadata', array( PodsInit::$meta, 'update_comment_meta' ), 10, 5 ),
-                array( 'delete_comment_metadata', array( PodsInit::$meta, 'delete_comment_meta' ), 10, 5 )
-            ) );
-        }
+			if ( !pods_tableless() ) {
+				$no_conflict[ 'filter' ] = array_merge( $no_conflict[ 'filter' ], array(
+					array( 'add_comment_metadata', array( PodsInit::$meta, 'add_comment_meta' ), 10, 5 ),
+					array( 'update_comment_metadata', array( PodsInit::$meta, 'update_comment_meta' ), 10, 5 ),
+					array( 'delete_comment_metadata', array( PodsInit::$meta, 'delete_comment_meta' ), 10, 5 )
+				) );
+			}
+		}
 
         $no_conflict[ 'action' ] = array(
             array( 'pre_comment_approved', array( PodsInit::$meta, 'validate_comment' ), 10, 2 ),
